converter logging added
also converter proj reorg
