+++ ConnectorAutocadCivil/ConnectorCivil2021/ConnectorCivil2021.csproj
@@ -71,13 +71,13 @@
   </ItemGroup>
   <ItemGroup>
     <PackageReference Include="Avalonia">
-      <Version>0.10.7</Version>
+      <Version>0.10.8</Version>
     </PackageReference>
     <PackageReference Include="Avalonia.Desktop">
-      <Version>0.10.7</Version>
+      <Version>0.10.8</Version>
     </PackageReference>
     <PackageReference Include="Avalonia.ReactiveUI">
-      <Version>0.10.7</Version>
+      <Version>0.10.8</Version>
     </PackageReference>
     <PackageReference Include="MSBuild.AssemblyVersion">
       <Version>1.3.0</Version>
